Update djd-site-post.php
Trying to get the confirmation email to mention which type of new post this submission is for.  Currently they all just say "Yocto Project".  Jefro needs the email to be specific.  I think this was just an oversite from Noel and a QA miss on our part.  Hopefully this VAR change will do what we need.